Remove issue and PR templates
Because those templates are in the "organization configuration" repo,
all repos within the vuestorefront github org which don't have an
issue template, inherit the issue from here.

This is particularly annoying in private repos where such detailed
issue templates are not necessary, as we work with eachother everyday
and know the tasks that we're developing.